initial commit to github
still not functional enough to call an initial version, but finally puts the code behind some version control.

as of this point:
   character classes are mostly complete as entities
      - can blend race/class to get stats and traits
      - currently outsourcing the attack details
   creature classes are mostly complete as entities
      - can create from dnd5e creature stat block info
      - currently extending to use attack functions
   most enum types necessary have init implementation
   conceptual framework for encounters is implemented
       - not ready to call initial simple build yet

current focus is on getting encounters up and running, implementing anything they need to reach the basic melee attack/defence capability given what is currently built.

in the midst of a minor refactor to better isolate this from being partially embedded within characters and creatures.